backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
Pillow>=10.0.0
python-docx>=0.8.11
pypdf>=3.0.0

# Google Gemini
google-generativeai>=0.7.2

# MetaTrader 5 integration
MetaTrader5>=5.0.45

# SMS with Kavenegar
kavenegar>=1.1.2

# Google OAuth
google-auth>=2.23.0
google-auth-httplib2>=0.1.1

# Zarinpal Payment Gateway
requests>=2.31.0

# ML/DL Optimization (optional - for advanced optimization)
# optuna>=3.0.0  # Uncomment if you want Bayesian optimization
# torch>=2.0.0  # Uncomment if you want Deep Learning features

